♻️  (vercel) NICE-15 analytics via app/dir
Was not tracking because I think it was in a client component not server in root fo `app/layout`
