+++ sites/jeromefitzgerald.com/src/components/Analytics/Vercel.tsx
@@ -1,0 +1,7 @@
+import { Analytics } from '@vercel/analytics/react'
+
+function VercelAnalytics() {
+  return <Analytics debug={process.env.NODE_ENV === 'production'} />
+}
+
+export { VercelAnalytics }
