+++ sites/jeromefitzgerald.com/src/components/Analytics/index.ts
@@ -1,1 +1,3 @@
 export { Analytics } from './Analytics'
+export { FathomAnalytics } from './Fathom'
+export { VercelAnalytics } from './Vercel'
